# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /collections:
    get:
      operationId: getCollectionByDateAndParlor
      tags: [collections]
      summary: Get the collection for a specific date and parlor
      security: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: parlorCode
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection found (or null)
          content:
            application/json:
              schema:
                type: object
                properties:
                  collection:
                    oneOf:
                      - $ref: "#/components/schemas/Collection"
                      - type: "null"
                required: [collection]
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a new collection entry
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "409":
          description: Already exists for this date and parlor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collections/{id}:
    get:
      operationId: getCollectionById
      tags: [collections]
      summary: Get a single collection by id
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCollection
      tags: [collections]
      summary: Update an existing draft collection
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "409":
          description: Cannot edit a submitted or acknowledged collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collections/{id}/submit:
    post:
      operationId: submitCollection
      tags: [collections]
      summary: Submit a collection to the supervisor
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "409":
          description: Already submitted or acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collections/{id}/acknowledge:
    post:
      operationId: acknowledgeCollection
      tags: [collections]
      summary: Supervisor acknowledges a submitted collection
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcknowledgeCollectionInput"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "409":
          description: Only submitted collections can be acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collections/list:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List collections for a date, scoped to the caller's access
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: agentCode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      $ref: "#/components/schemas/Collection"
                required: [collections]
  /collections/reports:
    get:
      operationId: getCollectionsReports
      tags: [collections]
      summary: List collections within a date range, scoped to the caller's access
      parameters:
        - name: dateFrom
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: agentCode
          in: query
          schema:
            type: string
        - name: parlorCode
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CollectionStatus"
      responses:
        "200":
          description: Collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      $ref: "#/components/schemas/Collection"
                required: [collections]
  /collections/supervisor:
    get:
      operationId: getSupervisorCollections
      tags: [collections]
      summary: List submitted/acknowledged collections for the supervisor's agents
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      $ref: "#/components/schemas/Collection"
                required: [collections]
...
  /external/parlor-summary/{parlorCode}/{date}:
    get:
      operationId: getExternalParlorSummary
      tags: [external]
      summary: Get the external system's collection summary for a parlor and date
      security: []
      parameters:
        - name: parlorCode
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: External summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalSummary"
        "503":
          description: External collection source is not configured or misconfigured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: External collection source is unavailable or returned invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          description: A message string, or a list of validation issues
      required: [error]
...
    CollectionStatus:
      type: string
      enum: [entered, submitted, acknowledged]
...
    AcknowledgeCollectionInput:
      type: object
      properties:
        acknowledgedBy:
          type: string
...
    Collection:
      type: object
      properties:
        id:
          type: integer
        parlorCode:
          type: string
        parlorName:
          type: string
        parlorType:
          type: string
        routeCode:
          type: string
        agentCode:
          type: string
        agentName:
          type: string
        collectionDate:
          type: string
          format: date
        cashAmount:
          type: string
          description: Decimal amount as a numeric string
        couponAmount:
          type: string
          description: Decimal amount as a numeric string
        ccAmount:
          type: string
          description: Decimal amount as a numeric string
        notes:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/CollectionStatus"
        submittedAt:
          type: string
          nullable: true
        acknowledgedAt:
          type: string
          nullable: true
        acknowledgedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
      required:
        - id
        - parlorCode
        - parlorName
        - parlorType
        - routeCode
        - agentCode
        - agentName
        - collectionDate
        - cashAmount
        - couponAmount
        - ccAmount
        - status
    CollectionInput:
      type: object
      properties:
        parlorCode:
          type: string
        parlorName:
          type: string
        parlorType:
          type: string
        routeCode:
          type: string
        agentCode:
          type: string
        agentName:
          type: string
        collectionDate:
          type: string
          format: date
        cashAmount:
          type: number
        couponAmount:
          type: number
        ccAmount:
          type: number
        notes:
          type: string
        status:
          $ref: "#/components/schemas/CollectionStatus"
      required:
        - parlorCode
        - parlorName
        - parlorType
        - routeCode
        - agentCode
        - agentName
        - collectionDate
    CollectionUpdateInput:
      type: object
      properties:
        cashAmount:
          type: number
        couponAmount:
          type: number
        ccAmount:
          type: number
        notes:
          type: string
        status:
          $ref: "#/components/schemas/CollectionStatus"
        submittedAt:
          type: string
        acknowledgedAt:
          type: string
        acknowledgedBy:
          type: string
...
    ExternalSummary:
      type: object
      properties:
        cashAmount:
          type: number
        couponAmount:
          type: number
        ccAmount:
          type: number
        source:
          type: string
        fetchedAt:
          type: string
      required: [cashAmount, couponAmount, ccAmount, source, fetchedAt]