# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionStatus } from './collectionStatus';

export interface Collection {
  id: number;
  parlorCode: string;
  parlorName: string;
  parlorType: string;
  routeCode: string;
  agentCode: string;
  agentName: string;
  collectionDate: Date;
  /** Decimal amount as a numeric string */
  cashAmount: string;
  /** Decimal amount as a numeric string */
  couponAmount: string;
  /** Decimal amount as a numeric string */
  ccAmount: string;
  notes?: string | null;
  status: CollectionStatus;
  submittedAt?: string | null;
  acknowledgedAt?: string | null;
  acknowledgedBy?: string | null;
  createdAt?: string | null;
  updatedAt?: string | null;
}