# lib/api-zod/src/generated/types/externalSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExternalSummary {
  cashAmount: number;
  couponAmount: number;
  ccAmount: number;
  source: string;
  fetchedAt: string;
}